Added Exception factory pipe for error handling
