+++ apps/official-journal-application-api/src/main.ts
@@ -7,16 +7,13 @@ import { WinstonModule } from 'nest-winston'
 import { apmInit } from '@dmr.is/apm'
 import { logger } from '@dmr.is/logging'
 
-import {
-  BadRequestException,
-  ValidationPipe,
-  VersioningType,
-} from '@nestjs/common'
+import { VersioningType } from '@nestjs/common'
 import { NestFactory } from '@nestjs/core'
 import { SwaggerModule } from '@nestjs/swagger'
 
 import { AppModule } from './app/app.module'
 import { openApi } from './openApi'
+import { ExceptionFactoryPipe } from '@dmr.is/pipelines'
 
 async function bootstrap() {
   const globalPrefix = 'api'
@@ -26,30 +23,7 @@ async function bootstrap() {
     logger: WinstonModule.createLogger({ instance: logger }),
   })
 
-  app.useGlobalPipes(
-    new ValidationPipe({
-      exceptionFactory(errors) {
-        const errs = errors.map((error) => {
-          const target = error.target?.constructor.name
-          logger.warn(
-            `Application API validation error: ${target}.${error.property} received<${error.value}>`,
-            {
-              constraints: error.constraints,
-              children: error.children,
-            },
-          )
-
-          return {
-            property: error.property,
-            constraints: error.constraints,
-          }
-        })
-        return new BadRequestException(errs)
-      },
-      transform: true,
-      // stopAtFirstError: true,
-    }),
-  )
+  app.useGlobalPipes(ExceptionFactoryPipe())
   app.setGlobalPrefix(globalPrefix)
   app.enableCors()
   app.enableVersioning({
