+++ libs/shared/pipelines/src/lib/index.ts
@@ -2,3 +2,4 @@ export * from './uuid.pipe'
 export * from './file-type.pipe'
 export * from './enum-type.pipe'
 export * from './is-string.pipe'
+export * from './exception-factory.pipe'
